+++ app/src/app/app.component.ts
@@ -1,0 +1,44 @@
+import { Component, OnInit } from '@angular/core';
+import { Title } from '@angular/platform-browser';
+import { Router, NavigationEnd, ActivatedRoute } from '@angular/router';
+
+import 'rxjs/add/operator/filter';
+import 'rxjs/add/operator/map';
+import 'rxjs/add/operator/mergeMap';
+
+import { config } from './config';
+
+@Component({
+    selector: 'uber-like',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss']
+})
+export class AppComponent implements OnInit {
+
+    constructor(
+        private router: Router,
+        private activatedRoute: ActivatedRoute,
+        private titleService: Title
+    ) { }
+
+    ngOnInit(): void {
+        /**
+         * We'll use reactive programming to set our title documents
+         */
+        this.router.events
+            .filter((event) => event instanceof NavigationEnd)
+            .map(() => this.activatedRoute)
+            .map((route) => {
+                while (route.firstChild) {
+                    route = route.firstChild;
+                }
+                return route;
+            })
+            .filter((route) => route.outlet === 'primary')
+            .mergeMap((route) => route.data)
+            .subscribe((event) => {
+                this.titleService.setTitle(event['title'] + config.suffixTitle);
+            });
+    }
+
+}