+++ app/src/app/config.ts
@@ -1,0 +1,3 @@
+export const config = {
+    suffixTitle: ' - Uber Like'
+};